refactor common test code to reduce duplication
