+++ searchlib/src/vespa/searchlib/fef/test/labels.cpp
@@ -1,0 +1,3 @@
+// Copyright Verizon Media. Licensed under the terms of the Apache 2.0 license. See LICENSE in the project root.
+
+#include "labels.h"
